# internal/crawler/infrastructure/repository/postgres/entities.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package postgres

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Source struct {
	ID      uuid.UUID        `json:"id"`
	Name    string           `json:"name"`
	Data    []byte           `json:"data"`
	Created pgtype.Timestamp `json:"created"`
	Updated pgtype.Timestamp `json:"updated"`
}